+++ main.go
@@ -63,14 +63,13 @@ func main() {
 
 	utils.SendRequestToWebhook(&configuration, newLog, "scan.started", []byte(`{"reason":"Scan Started successfully"}`))
 
+	defer utils.SendCompleteScanRequest(&configuration, newLog)
+
 	err = StartScansInRoutine(&configuration)
 
 	if err != nil {
 		newLog.Panicf("Error occurred while running scans -> %s", err.Error())
 		log.Exit(1)
 	}
-
 	newLog.Info("All scans completed successfully exiting")
-
-	utils.SendCompleteScanRequest(&configuration, newLog)
 }